openai-translator/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
pandas>=2.0.0
Pillow>=10.0.0
simplejson
requests
PyYAML
loguru
tabulate